Reorganize functions files
