+++ includes/functions-comment.php
@@ -3,6 +3,11 @@
 namespace ucare;
 
 
+// Update the comment status after a comment has been made
+add_action( 'comment_post', 'ucare\comment_save' );
+add_action( 'edit_comment', 'ucare\comment_save' );
+
+
 function comment_save($id ) {
 
     $post = get_post( get_comment( $id )->comment_post_ID );
@@ -29,7 +34,3 @@ function comment_save($id ) {
     }
 
 }
-
-// Update the comment status after a comment has been made
-add_action( 'comment_post', 'ucare\comment_save' );
-add_action( 'edit_comment', 'ucare\comment_save' );
