+++ includes/functions-deprecated.php
@@ -1,0 +1,3 @@
+<?php
+
+namespace ucare;
