+++ includes/trait-data.php
@@ -4,7 +4,6 @@
  * @package ucare
  * @since 1.4.2
  */
-
 namespace ucare;
